refactor: add buildSpanLink function to share createSpanLink procedure
